requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
apify_client==1.8.1
apify_shared==1.1.2
attrs==24.2.0
boto3==1.35.68
botocore==1.35.68
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
couchbase==4.3.3
fastapi==0.115.5
google_search_results==2.4.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jmespath==1.0.1
lxml==5.3.0
more-itertools==10.5.0
networkx==3.4.2
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pydantic==2.10.1
pydantic_core==2.27.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrends==4.9.2
pytz==2024.2
requests==2.32.3
s3transfer==0.10.4
selenium==4.27.1
serpapi==0.1.5
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.41.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
